# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripzant API Documentation",
        "description": "API endpoints for Tripzant Mobile App",
        "contact": {
            "email": "support@tripzant.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://tripnstays.com/api/v1",
            "description": "Staging/Live Server"
        }
    ],
    "paths": {
...
        "/get-exchange-rate": {
            "get": {
                "tags": [
                    "Money Transfer"
                ],
                "summary": "Get Exchange Rate",
                "operationId": "629ac3a2b2c228120bd0883c1f968df0",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "INR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange rate result"
                    }
                }
            }
        },
        "/get-providers": {
            "get": {
                "tags": [
                    "Money Transfer"
                ],
                "summary": "Get Transfer Providers",
                "operationId": "a2cfe42c51ef981cd90b25b4a0e54ba8",
                "parameters": [
                    {
                        "name": "from_country",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "US"
                        }
                    },
                    {
                        "name": "from_currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to_country",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "IN"
                        }
                    },
                    {
                        "name": "to_currency",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "INR"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of providers"
                    }
                }
            }
        },
        "/create-transfer": {
            "post": {
                "tags": [
                    "Money Transfer"
                ],
                "summary": "Create a Money Transfer",
                "operationId": "072ec1b04a3e851c68fdd19e55b053f7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "type",
                                    "source_currency",
                                    "source_amount",
                                    "target_currency",
                                    "target_amount",
                                    "exchange_rate",
                                    "recipient_details"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "wallet_to_wallet",
                                            "wallet_to_bank",
                                            "international"
                                        ]
                                    },
                                    "source_currency": {
                                        "type": "string"
                                    },
                                    "source_amount": {
                                        "type": "number"
                                    },
                                    "target_currency": {
                                        "type": "string"
                                    },
                                    "target_amount": {
                                        "type": "number"
                                    },
                                    "exchange_rate": {
                                        "type": "number"
                                    },
                                    "recipient_details": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer created successfully"
                    }
                }
            }
        },
...
    },
...
}